# docs/openapi.yaml
openapi: 3.1.0
info:
  title: London Property Explorer API
  version: 1.0.0
paths:
  /api/health:
    get:
      tags:
      - system
      summary: Health
      operationId: health_api_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /api/meta:
    get:
      tags:
      - system
      summary: Meta
      operationId: meta_api_meta_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaResponse'
  /api/capabilities:
    get:
      tags:
      - system
      summary: Capabilities
      operationId: capabilities_api_capabilities_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
  /api/transactions:
    get:
      tags:
      - data
      summary: Transactions
      operationId: transactions_api_transactions_get
      parameters:
      - name: bbox
        in: query
        required: true
        schema:
          type: string
          title: Bbox
      - name: zoom
        in: query
        required: true
        schema:
          type: integer
          maximum: 22
          minimum: 0
          title: Zoom
      - name: min_price
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 50000000
            minimum: 0
          - type: 'null'
          title: Min Price
      - name: max_price
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 50000000
            minimum: 0
          - type: 'null'
          title: Max Price
      - name: types
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Types
      - name: tenures
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenures
      - name: from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: From
      - name: to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: To
      - name: format
        in: query
        required: false
        schema:
          enum:
          - json
          - bin
          type: string
          default: json
          title: Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/districts:
    get:
      tags:
      - data
      summary: Districts
      operationId: districts_api_districts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/district-stats:
    get:
      tags:
      - data
      summary: District Stats
      operationId: district_stats_api_district_stats_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DistrictStats'
                type: array
                title: Response District Stats Api District Stats Get
  /api/postcode/{postcode}/history:
    get:
      tags:
      - data
      summary: Postcode History
      operationId: postcode_history_api_postcode__postcode__history_get
      parameters:
      - name: postcode
        in: path
        required: true
        schema:
          type: string
          title: Postcode
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostcodeHistory'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/chat:
    post:
      tags:
      - assistant
      summary: Chat
      operationId: chat_api_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/chat/stream:
    post:
      tags:
      - assistant
      summary: Chat Stream
      operationId: chat_stream_api_chat_stream_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/chat/{run_id}/feedback:
    post:
      tags:
      - assistant
      summary: Feedback
      operationId: feedback_api_chat__run_id__feedback_post
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AgentMetrics:
      properties:
        route:
          type: string
          enum:
          - sql
          - rag
          - hybrid
          - map_action
          - unsupported
          title: Route
        latency_ms:
          type: integer
          minimum: 0.0
          title: Latency Ms
        input_tokens:
          type: integer
          minimum: 0.0
          title: Input Tokens
        output_tokens:
          type: integer
          minimum: 0.0
          title: Output Tokens
        estimated_cost_usd:
          type: number
          minimum: 0.0
          title: Estimated Cost Usd
        graph_version:
          type: string
          title: Graph Version
        prompt_hash:
          type: string
          title: Prompt Hash
        model:
          type: string
          title: Model
        corpus_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Corpus Version
      additionalProperties: false
      type: object
      required:
      - route
      - latency_ms
      - input_tokens
      - output_tokens
      - estimated_cost_usd
      - graph_version
      - prompt_hash
      - model
      - corpus_version
      title: AgentMetrics
    CapabilityResponse:
      properties:
        chat:
          type: boolean
          title: Chat
        rag:
          type: boolean
          title: Rag
        tracing:
          type: boolean
          title: Tracing
        streaming:
          type: boolean
          title: Streaming
        feedback:
          type: boolean
          title: Feedback
        graph_version:
          type: string
          title: Graph Version
        corpus_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Corpus Version
      additionalProperties: false
      type: object
      required:
      - chat
      - rag
      - tracing
      - streaming
      - feedback
      - graph_version
      - corpus_version
      title: CapabilityResponse
    ChatMessage:
      properties:
        role:
          type: string
          enum:
          - user
          - assistant
          title: Role
        content:
          type: string
          maxLength: 2000
          minLength: 1
          title: Content
      additionalProperties: false
      type: object
      required:
      - role
      - content
      title: ChatMessage
    ChatRequest:
      properties:
        messages:
          items:
            $ref: '#/components/schemas/ChatMessage'
          type: array
          maxItems: 12
          minItems: 1
          title: Messages
      additionalProperties: false
      type: object
      required:
      - messages
      title: ChatRequest
    ChatResponse:
      properties:
        run_id:
          type: string
          format: uuid
          title: Run Id
        reply:
          type: string
          maxLength: 8000
          minLength: 1
          title: Reply
        citations:
          items:
            $ref: '#/components/schemas/Citation'
          type: array
          maxItems: 10
          title: Citations
        steps:
          items:
            $ref: '#/components/schemas/ChatStep'
          type: array
          maxItems: 20
          title: Steps
        map_action:
          anyOf:
          - $ref: '#/components/schemas/MapAction'
          - type: 'null'
        degraded:
          type: boolean
          title: Degraded
          default: false
        metrics:
          $ref: '#/components/schemas/AgentMetrics'
      additionalProperties: false
      type: object
      required:
      - run_id
      - reply
      - citations
      - steps
      - metrics
      title: ChatResponse
    ChatStep:
      properties:
        name:
          type: string
          maxLength: 80
          minLength: 1
          title: Name
        status:
          type: string
          enum:
          - completed
          - degraded
          - failed
          title: Status
        detail:
          type: string
          maxLength: 300
          minLength: 1
          title: Detail
        duration_ms:
          type: integer
          minimum: 0.0
          title: Duration Ms
      additionalProperties: false
      type: object
      required:
      - name
      - status
      - detail
      - duration_ms
      title: ChatStep
    Citation:
      properties:
        id:
          type: string
          maxLength: 128
          minLength: 1
          title: Id
        title:
          type: string
          maxLength: 300
          minLength: 1
          title: Title
        section:
          anyOf:
          - type: string
            maxLength: 300
          - type: 'null'
          title: Section
        publisher:
          type: string
          maxLength: 200
          minLength: 1
          title: Publisher
        url:
          type: string
          pattern: ^https?://
          title: Url
        licence:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Licence
      additionalProperties: false
      type: object
      required:
      - id
      - title
      - publisher
      - url
      title: Citation
    DistrictStats:
      properties:
        district:
          type: string
          pattern: ^[A-Z]{1,2}[0-9][0-9A-Z]?$
          title: District
        sales:
          type: integer
          exclusiveMinimum: 0.0
          title: Sales
        median_price:
          type: integer
          maximum: 50000000.0
          minimum: 10000.0
          title: Median Price
      additionalProperties: false
      type: object
      required:
      - district
      - sales
      - median_price
      title: DistrictStats
    FeedbackRequest:
      properties:
        score:
          type: integer
          enum:
          - -1
          - 1
          title: Score
        reason:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Reason
        correction:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Correction
      additionalProperties: false
      type: object
      required:
      - score
      title: FeedbackRequest
    FeedbackResponse:
      properties:
        accepted:
          type: boolean
          title: Accepted
        trace_attached:
          type: boolean
          title: Trace Attached
      additionalProperties: false
      type: object
      required:
      - accepted
      - trace_attached
      title: FeedbackResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HealthResponse:
      properties:
        status:
          type: string
          const: ok
          title: Status
          default: ok
      additionalProperties: false
      type: object
      title: HealthResponse
    HistoryEntry:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        price:
          type: integer
          maximum: 50000000.0
          minimum: 10000.0
          title: Price
        date:
          type: string
          format: date
          title: Date
        type:
          type: string
          enum:
          - D
          - S
          - T
          - F
          - O
          title: Type
        tenure:
          type: string
          enum:
          - F
          - L
          title: Tenure
        is_new:
          type: boolean
          title: Is New
        paon:
          anyOf:
          - type: string
          - type: 'null'
          title: Paon
        saon:
          anyOf:
          - type: string
          - type: 'null'
          title: Saon
        street:
          anyOf:
          - type: string
          - type: 'null'
          title: Street
        town:
          anyOf:
          - type: string
          - type: 'null'
          title: Town
      additionalProperties: false
      type: object
      required:
      - id
      - price
      - date
      - type
      - tenure
      - is_new
      - paon
      - saon
      - street
      - town
      title: HistoryEntry
    MapAction:
      properties:
        kind:
          type: string
          enum:
          - set_filters
          - highlight_district
          title: Kind
        payload:
          additionalProperties: true
          type: object
          title: Payload
        label:
          type: string
          maxLength: 120
          minLength: 1
          title: Label
      additionalProperties: false
      type: object
      required:
      - kind
      - payload
      - label
      title: MapAction
    MetaResponse:
      properties:
        total:
          type: integer
          exclusiveMinimum: 0.0
          title: Total
        from:
          type: string
          format: date
          title: From
        to:
          type: string
          format: date
          title: To
      additionalProperties: false
      type: object
      required:
      - total
      - from
      - to
      title: MetaResponse
    PostcodeHistory:
      properties:
        postcode:
          type: string
          pattern: ^[A-Z]{1,2}[0-9][0-9A-Z]? [0-9][A-Z]{2}$
          title: Postcode
        count:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Count
        truncated:
          type: boolean
          title: Truncated
        entries:
          items:
            $ref: '#/components/schemas/HistoryEntry'
          type: array
          maxItems: 200
          minItems: 1
          title: Entries
      additionalProperties: false
      type: object
      required:
      - postcode
      - count
      - truncated
      - entries
      title: PostcodeHistory
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError